Make base64 decoder check RFC 4648 zero padding and reject invalid
In an external audit of our code, the following security concerns
were raised about the base64 decoding implementation:

> RFC4648 specifies that base64 encoding implementations may reject
> input that is not appropriately zero padded during decoding. The
> binascii crate’s base64 implementation does not appear to reject
> input strings that are not zero padded. Depending of the context
> in which it is used, it may open a service using this crate to
> attacks.

They then went on to demonstrate that because binascii doesn't do
this, the encoding is "malleable" in the sense that there are byte
sequences which are different, but decode to the same payload.
This is possible because one of the byte sequences may have invalid
zero padding.

In this commit we add checks in the decoder that validates the zero
padding. Then we check that expected invalid strings are rejected.
Then we check that their method of finding "collisions" in the encoding
doesn't produce any more collisions when we try it with a few
different strings.